Add a basic graph data structure
The implementation using matrices with floating-point values isn't the
most efficient one but it's definitely simple and allows me to reuse the
matrix implementation that I'll use for other demonstrations too. In
real life a matrix template (in C++) or matrix code generator (in C)
would be a better idea.